# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Retrieves concordances, frequency information and collocations from language corpora",
        "title": "MQuery - query and analyze corpus data",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/corplist": {
            "get": {
                "description": "Shows a list of available corpora with their basic properties.",
                "produces": [
                    "application/json"
                ],
                "summary": "Corplist",
                "parameters": [
                    {
                        "type": "string",
                        "default": "en",
                        "description": "An ISO 639-1 locale code of response",
                        "name": "lang",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Corplist"
                        }
                    }
                }
            }
        },
...
        "/info/{corpusId}": {
            "get": {
                "description": "Shows a detailed corpus information, including size in tokens, available positional and structural attributes.",
                "produces": [
                    "application/json"
                ],
                "summary": "CorpusInfo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "An ID of a corpus to get info about",
                        "name": "corpusId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "en",
                        "description": "An ISO 639-1 locale code of response.",
                        "name": "locale",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/CorpusInfo"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Citation": {
            "type": "object",
            "properties": {
                "article_ref": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "default_ref": {
                    "type": "string"
                },
                "other_bibliography": {
                    "type": "string"
                }
            }
        },
...
        "Corplist": {
            "type": "object",
            "properties": {
                "corpora": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/CorpusCompactInfo"
                    }
                },
                "locale": {
                    "type": "string"
                }
            }
        },
        "Corpus": {
            "type": "object",
            "properties": {
                "attrList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Item"
                    }
                },
                "citationInfo": {
                    "$ref": "#/definitions/Citation"
                },
                "corpname": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "srchKeywords": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "structList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Item"
                    }
                },
                "textProperties": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "webUrl": {
                    "type": "string"
                }
            }
        },
        "CorpusCompactInfo": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "flags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "subcorpora": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/SubcInfo"
                    }
                }
            }
        },
        "CorpusInfo": {
            "type": "object",
            "properties": {
                "corpus": {
                    "$ref": "#/definitions/results.CorpusInfo"
                },
                "locale": {
                    "type": "string"
                }
            }
        },
...
        "Item": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
...
        "SubcInfo": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
        "results.CorpusInfo": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/Corpus"
                },
                "error": {}
            }
        },
...
    },
...
}